Remove a circular dependency
